Add oneOf/anyOf/allOf and response code check
This commit adds support for OpenAPI's oneOf, anyOf, and allOf keywords.
This is accomplished through the creation of the InPoly class and its
subclasses OneOf, AnyOf, and AllOf. These classes all contain a dump
method and are callable, returing a Schema instance as is expected by
the webargs use_args method. This commit also adds a check to prevent
registration of a status code to a view method more than once.